pyproject.toml
[tool.poetry]
name = "numbering"
version = "0.1.0"
description = ""
authors = ["Madung2 <tuliphan91@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-docx = "^1.1.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
